requirements.txt
ogb==1.3.0
numpy==1.19.2
torch==1.7.1
dgl_cu110==0.5.3
